+++ src/Publisher.cc
@@ -115,7 +115,7 @@ size_t Publisher::Pack(char *_buffer) const
   }
 
   // Pack the topic length.
-  uint16_t topicLength = this->topic.size();
+  uint16_t topicLength = static_cast<uint16_t>(this->topic.size());
   memcpy(_buffer, &topicLength, sizeof(topicLength));
   _buffer += sizeof(topicLength);
